Enable ccache on the CI (#435)
This change makes use of ccache the default setting for rocFFT builds on
the CI. It can be disabled for PRs using the label 'ci:no-ccache'.

I've also removed the use of sudo from the build commands, as sudo is
no longer necessary for building on any CI image. It must be removed,
as the sudo command undoes the PATH modifications used to enable ccache.