fix padding on 2D R2C with nested L1D_TRTRT
* padding: collect parent temp ops in the correct order

We were getting temp buf ops in the wrong order if the plan looked like:

  1a: B -> T
    2a: B -> B
    2b: B -> T
      3a: B -> T
      3b: T -> B
      3c: B -> T
    2c: T -> T
  1b: T -> B

The temp buffer usage goes from 3c -> 2b -> 2c -> 1a -> 1b, in that
order.  But we were traversing the tree structure in the wrong order
and mixed that up.

* padding: skip in-place R2C/C2R

These nodes change the data length and confuse padding.  So just skip
those cases.

* rocfft-test: re-enable adhoc tests that had broken padding