improve strided L1D_CC performance
* generator: allow using batch for largeTwd transform count

Requires RTC if enabled.

* tree_node: create transpose nodes with correct length/stride

Make them correct to begin with, instead of making them wrong and fixing
them up during buffer assignment.

* assignment_policy: use ptrdiff to decide validity

This fixes assignment checking when in/out buffers have unusual strides.

* tree_node: special case strided L1D_CC plans

L1D_CC normally assumes the sub-dimensional columns are adjacent in
memory.  In the strided case, dist is 1 and the batch dimension is what's
adjacent.  So use batch as the adjacent dimension in that case to improve
memory access coalescing.