Use feature check macro for OpenMP (#984)
In ROCm, OpenMP support is provided by the rocm-openmp-extras package,
which is somewhat complex and is therefore not always available on
platforms that lack official support from AMD.

The OpenMP standard requires compilers that support OpenMP to define
the _OPENMP macro as the OpenMP standard version in yyyymm format.
These checks match those done in rocSPARSE to make OpenMP an optional
dependency.

However, the _OPENMP macro does not seem to be defined when building
rocFFT with OpenMP. It's unclear whether that is due to a defect in
the OpenMP implementation or the rocFFT build system. Either way, we
can work around the problem by using a manually defined macro.