Use the standard _OPENMP macro (#1063)
The _OPENMP macro is the standard way to check if OpenMP is available
at compile-time. I had been confused by the behaviour of the macro when
I authored e7b1fe244ab0623e900b4efe75be29545df1163b, but the problem
is now more clear. As per the HIP FAQ [1], the _OPENMP macro is not defined
during device compilation. That is fine, as long as the _OPENMP macro
is used consistently to guard all uses of OpenMP.

[1]: https://github.com/ROCm-Developer-Tools/HIP/blob/rocm-5.4.4/docs/markdown/hip_faq.md#why-_openmp-is-undefined-when-compiling-with--fopenmp