rocfft-test: don't update cached FFT batch count if only batch got smaller
If tests are ordered like:

  case 1: length N double precision batch 4
  case 2: length N double precision batch 2
  case 3: length N double precision batch 1
  case 4: length N single precision batch 4
  case 5: length N single precision batch 2
  case 6: length N single precision batch 1

We were incorrectly updating the batch count of cached CPU data on each
test.  So at case 4, we would need to rerun the CPU FFT because the
stored batch count was only 1, even though we are still keeping around
batch-4 data computed in case 1.

Fix this by remembering that the cached data really is batch-4.