move rtc helper, system cache to subdir
* rtc_cache: move system cache to rocfft subdir, next to lib

Search for system cache both directly adjacent to lib, and in subdir.
In build directories, it'll be beside the lib, but in installs, it'll
be in the subdir.

* rocfft_rtc_helper: move to rocfft subdir on Linux

Architecture-dependent internal executables should go into a rocfft
subdir of lib.  On Windows, keep the helper next to the lib to simplify
DLL loading.